+++ crates/core/src/world/generate/elevation.rs
@@ -30,6 +30,9 @@ impl Generate for ElevationGenerator {
             let pos = tile.position();
             let d = pos.distance_to(HexPoint::ORIGIN) as f64;
 
+            // Determine the range of potential elevation outputs for this tile.
+            // For most tiles it's static, but for some the edge buffer will
+            // restrict that range
             let elev_range: NumRange<Meter, f64> = if buffer_range.contains(d) {
                 // This tile is near the edge of the world, so we want to push
                 // it down a bit. The further out it is, the more we push it.
@@ -39,13 +42,19 @@ impl Generate for ElevationGenerator {
                 // normal. At the outermost ring, it'll be sea level. This
                 // guarantees at least one ring of ocean at the edge.
                 let elev_max = buffer_range
+                    // Convert the value to a fraction representing its distance
+                    // from the outermost edge. 0 will be the outermost ring,
+                    // 1 will be the innermost ring **of the buffer**
                     .value(d)
                     .normalize()
                     .invert()
-                    // We now have a value where 0 is the outermost ring and 1
-                    // is the innermost ring OF THE BUFFER
+                    // Apply exponent curve
                     .apply(|v| v.powf(config.edge_buffer_exponent)) // Use a smooth gradient
                     .convert::<Meter>()
+                    // Pick a new upper bound on elevation, somewhere between
+                    // sea level and the standard upper bound. For the
+                    // outermost ring, this will be sea level, innermost will
+                    // remain the standard value
                     .map_to(NumRange::new(
                         World::SEA_LEVEL,
                         World::ELEVATION_RANGE.max,
@@ -58,8 +67,12 @@ impl Generate for ElevationGenerator {
                 World::ELEVATION_RANGE
             };
 
+            // This raw value will be [-1,1] (ish).
+            // TODO https://github.com/LucasPickering/terra-rs/issues/19
+            // Figure out why values aren't spanning the full [-1,1] range
+            let raw_noise = noise_fn.get(pos);
             let elev = normal_range
-                .value(noise_fn.get(pos))
+                .value(raw_noise)
                 .convert::<Meter>()
                 .map_to(elev_range)
                 .inner();
