+++ crates/core/tests/test_config_validation.rs
